% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/debugNextItem.R
\name{debugNextItem}
\alias{debugNextItem}
\title{Computerized Adaptive Testing Survey Debug Next Item Function}
\usage{
debugNextItem(cat, theta.est = NA, D = 1, lowerBound = -4,
  upperBound = 4, quadPoints = 33)
}
\arguments{
\item{cat}{An object of class \code{CATsurv}}

\item{theta.est}{A scalar value to contain an estimate of a respondent's position on the latent trait, using the \code{\link{estimateTheta}} funciton. Defaults to NA.}

\item{D}{A numeric value used as model parameter.  For logit models, set D=1.  For an approximation of the probit model, set D=1.702.  Defaults to D=1.}

\item{lowerBound}{The lower bound of the interval of the latent trait used in estimation.  Defaults to -4.}

\item{upperBound}{The upper bound of the interval of the latent trait used in estimation.  Defaults to 4.}

\item{quadPoints}{The number of points used in approximating the integral.  Defaults to 33.}

\item{D}{A numeric value used as model parameter.  For logit models, set D=1.  For an approximation of the probit model, set D=1.702.  Defaults to D=1.}

\item{answer}{The answer to the item \emph{k} to be stored}
}
\value{
A next item to be asked based on the estimate of a respondent's position on the latent trait
}
\description{
This function enables user to debug a next item based on an estimate of a respondent's position by plotting difficulty of next item with respect to its expected posterior variance.
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{three.pl}},\code{\link{likelihood}}, \code{\link{prior.value}}, \code{\link{estimateTheta}}, \code{\link{estimateSE}}, \code{\link{expectedPV}},  \code{\link{nextItem}}, \code{\link{storeAnswer}}
}

