% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimateSE.R
\name{estimateSE}
\alias{estimateSE}
\title{Computerized Adaptive Testing Survey Posterior Standard Error of Estimated Latent Trait Position Estimator}
\usage{
estimateSE(cat, theta.hat, ...)
}
\arguments{
\item{cat}{An object of class \code{CATsurv}}

\item{theta.hat}{A scalar value containing an estimate of a respondent's position on the latent trait.  Generally, this is the output of the \code{\link{estimateTheta}} funciton.}

\item{...}{argument passed to other functions.}
}
\value{
The estimate of the standard error of the user supplied \code{theta.hat}
}
\description{
This function estimates the standard error of an estimate of a respondent's expected \emph{a posteriori} (EAP) position on the latent scale.
}
\details{
The standard error of the expected \emph{a posteriori} (EAP) estimate of respondent \eqn{j}'s position on the latent scale is calculated as the square root of \deqn{E((\theta_j-\hat{\theta_j}^{(\text{EAP})})^2|\mathbf{y}_{k-1,j})=\frac{\int(\theta_j-\hat{\theta_j}^{(\text{EAP})})^2\pi(\theta_j)L(\theta_j|\mathbf{y}_{k-1,j}d\theta_j}{\int\pi(\theta_j)L(\theta_j|\mathbf{y}_{k-1,j})d\theta_j}}.
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{three.pl}},\code{\link{likelihood}}, \code{\link{prior.value}}, \code{\link{estimateTheta}}, \code{\link{expectedPV}}, \code{\link{nextItem}}, \code{\link{storeAnswer}}, \code{\link{debugNextItem}}
}

