% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimateThetaML.R
\name{estimateThetaML}
\alias{estimateThetaML}
\title{Computerized Adaptive Testing Survey Estimating Each Respondent's Position on the Latent Scale of Interest - Maximum Likelihood method}
\usage{
estimateThetaML(cat, ...)
}
\arguments{
\item{cat}{an object of class \code{CATsurv}.}

\item{...}{arguement passed to other functions}
}
\value{
A scalar with the ML estimate of the respondent's place on the latent scale
}
\description{
This function returns the maximum likelihood (ML) estimate of each respondent's popsition on the latent scale.
}
\details{
The ML estimate is based off of the likelihood funciton \deqn{L(\theta|\mathbf{y}_{k-1}) = \prod_{i=1}^{k-1}p_i(\theta_j)^{y_ij}q_i(\theta_j)^{(1-y_{ij})}}.  This procedure does not produce finite estimates for response patters with all items correct or all items incorrect.
}
\author{
Josh W. Cutler and Jacob M. Montgomery
}
\seealso{
\code{\link{three.pl}},\code{\link{likelihood}}, \code{\link{prior.value}}, \code{\link{estimateSE}}, \code{\link{expectedPV}}, \code{\link{nextItem}}, \code{\link{storeAnswer}}, \code{\link{debugNextItem}}
}

