% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expectedPV.R
\name{expectedPV}
\alias{expectedPV}
\title{Computerized Adaptive Testing Survey Expected Posterior Variance Estimator}
\usage{
expectedPV(cat, item)
}
\arguments{
\item{cat}{an object of class \code{CATsurv}}

\item{item}{The question for which to estimate the expected posterior variance for a respondent with a latent trait estimate of theta.hat.  This should be the name of a row in the "questions" data-frame in the "questions" slot of a \code{CATsurv} object.}
}
\value{
The expected posterior variance for respondent \emph{j} on item \emph{k}.
}
\description{
This function estimates the expected posterior variance for a respondent's estimated position on the latent trait on an item which he/she has yet to answer based on a respondent's position on the latent trait from the items he/she has already answered.
}
\details{
The expected posterior variance is calculated as \deqn{P(y^*_{kj}=1|\mathbf{y}_{k-1,j})\text{Var}(\theta_j|\mathbf{y}_{k-1,j},y^*_{kj}=1)+P(y^*_{kj}=0|\mathbf{y}_{k-1,j})\text{Var}(\theta_j|\mathbf{y}_{k-1,j},y^*_{kj}=0)}, where \eqn{y_{kj}^*} is a possible response to item \emph{k} by respondent \emph{j}.
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{three.pl}},\code{\link{likelihood}}, \code{\link{prior.value}}, \code{\link{estimateTheta}}, \code{\link{estimateSE}}, \code{\link{nextItem}}, \code{\link{storeAnswer}}, \code{\link{debugNextItem}}
}

