% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nextItem.R
\name{nextItem}
\alias{nextItem}
\title{Computerized Adaptive Testing Survey Next Item Function}
\usage{
nextItem(cat, ability.estimator = "EAP", item.selection = "EPV")
}
\arguments{
\item{cat}{An object of class \code{CATsurv}}

\item{ability.estimator}{The estimation procedure used to estimate the respondent's position on the latent scale.  The three options are "EAP" for expected a posterior (the default),  "ML" for maximum likelihood, and "MAP" for maximum a posterior.}

\item{item.selection}{The item selection procedure.  The five options are "EPV" for minimum expected posterior variance (the default),  "KL" for Kullback-Leibler, "MFI" for maximum Fisher's information, "MPWI" for maximum posterior wieghted information, "MWFI" for maximum weighted Fisher's information, "MEI" for maximum expected information.}
}
\value{
A data frame of available questions based on the use selected item selection criterion for the respondent and a row name for the next item to be asked
}
\description{
This function takes a respondent's previous answers to determine the next item from the list of available questions determined according to a user specified methods for latent trait estimation and item selection.
}
\author{
Josh W. Cutler: \email{josh@zistle.com} and Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
\seealso{
\code{\link{three.pl}},\code{\link{likelihood}}, \code{\link{prior.value}}, \code{\link{estimateTheta}}, \code{\link{estimateSE}}, \code{\link{expectedPV}},  \code{\link{storeAnswer}}, \code{\link{debugNextItem}}
}

