% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nextItemMFI.R
\name{nextItemMFI}
\alias{nextItemMFI}
\title{Computerized Adaptive Testing Survey Maximum Fisher's Information Function}
\usage{
nextItemMFI(cat)
}
\arguments{
\item{cat}{an object of \code{CATsurv} class.}
}
\value{
The next item to present to the respondent with the largest Fisher's information.
}
\description{
This function determines the next item by comparing Fisher's information and choosing the largest and presents it to the respondent.
}
\author{
Josh W. Cutler and Jacob M. Montgomery
}
\seealso{
\code{\link{likelihood}},\code{\link{prior}}, \code{\link{estimateTheta}}, \code{\link{estimateSE}}, \code{\link{expectedPV}}, \code{\link{nextItem}}, \code{\link{storeAnswer}}, \code{\link{debugNextItem}}
}

