% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\title{Computerized Adaptive Testing Survey Prior Value Function}
\usage{
prior(cat, values, name, params)
}
\arguments{
\item{cat}{an object of class \code{CATsurv}.}

\item{values}{vector consisting of quantile for each respondent.}

\item{name}{character string for type of assumed distribution. \code{normal} for normal distribution, \code{cauchy} for Cauchy distribtuion, and \code{t} for t-distribution.}

\item{params}{vector of parameters for selected distribution. The first element for mean value, the second element for standard deviation, and the third element for degree of freedom.}
}
\value{
A vector consisting of prior value for each respondent according to the assumed distribution.
}
\description{
This function returns the prior value for each repondent's position on the latent scale of interest.
}
\author{
Josh W. Cutler and Jacob M. Montgomery
}
\seealso{
\code{\link{three.pl}},\code{\link{likelihood}}, \code{\link{estimateTheta}}, \code{\link{estimateSE}}, \code{\link{expectedPV}}, \code{\link{nextItem}}, \code{\link{storeAnswer}}, \code{\link{debugNextItem}}
}

