% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/question.path.R
\name{question.path}
\alias{question.path}
\title{Computerized Adaptive Testing Survey Possible Combination of Question Paths}
\usage{
question.path(cat, ability.estimator = "EAP", item.selection = "EPV")
}
\arguments{
\item{cat}{an object of \code{CATsurv} class.}

\item{ability.estimator}{The estimation procedure used to estimate the respondent's position on the latent scale.  The three options are "EAP" for expected a posterior (the default),  "ML" for maximum likelihood, and "MAP" for maximum a posterior.}

\item{item.selection}{The item selection procedure.  The five options are "EPV" for minimum expected posterior variance (the default),  "KL" for Kullback-Leibler, "MFI" for maximum Fisher's information, "MPWI" for maximum posterior wieghted information, "MWFI" for maximum weighted Fisher's information, "MEI" for maximum expected information.}
}
\value{
A list consisting of the all possible question paths according to the answer at each stage.
}
\description{
This function returns the all possible combinations of question paths up to 5th stage or \code{n-1}th stage when the length of questions is less than or equal to 5.
}
\author{
Josh W. Cutler and Jacob M. Montgomery
}
\seealso{
\code{\link{likelihood}},\code{\link{prior.value}}, \code{\link{estimateTheta}}, \code{\link{estimateSE}}, \code{\link{expectedPV}}, \code{\link{nextItem}}, \code{\link{storeAnswer}}, \code{\link{debugNextItem}}
}

